presentation user module
