+++ server/kernel/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 
 [dependencies]
 
-diesel = { version = "1", features = ["postgres", "chrono"] }
+diesel = { version = "1", features = ["postgres", "chrono", "r2d2"] }
 diesel_migrations = "1"
 diesel-derive-newtype = "1"
